// backed/ecgbc-admin-backend/src/app/features/permission/enums/permission.enum.ts
export enum MemberPermission {
    MEMBER_ADD = "add_member",
    MEMBER_CHANGE = "change_member",
    MEMBER_VIEW = "view_member",
    MEMBER_DELETE = "delete_member",
    MEMBER_DEACTIVATE = "deactivate_member",
  }
  
  export enum StaffPermission {
    STAFF_ADD = "add_staff",
    STAFF_CHANGE = "change_staff",
    STAFF_VIEW = "view_staff",
    STAFF_DELETE = "delete_staff",
  }
  
  export enum PermissionPermission {
    PERMISSION_ADD = "add_permission",
    PERMISSION_CHANGE = "change_permission",
    PERMISSION_VIEW = "view_permission",
    PERMISSION_DELETE = "delete_permission",
  }
  
  export enum RolePermission {
    ROLE_ADD = "add_role",
    ROLE_CHANGE = "change_role",
    ROLE_VIEW = "view_role",
    ROLE_DELETE = "delete_role",
  }

  export enum CouncilFellowship {
    COUNCIL_FELLOWSHIP_ADD = "add_fellowship",
    COUNCIL_FELLOWSHIP_CHANGE = "change_fellowship",
    COUNCIL_FELLOWSHIP_VIEW = "view_fellowship",
    COUNCIL_FELLOWSHIP_DELETE = "delete_fellowship",
    COUNCIL_FELLOWSHIP_DEACTIVATE = "deactivate_fellowship",
  }

  export enum FilePermission  {
    FILE_ADD = "add_file",
    FILE_CHANGE = "change_file",
    FILE_VIEW = "view_file",
    FILE_DELETE = "delete_file"
  }
  export enum ReportPermission  {
    REPORT_ADD = "add_report",
    REPORT_CHANGE = "change_report",
    REPORT_VIEW = "view_report",
    REPORT_DELETE = "delete_report"
  }

 